Added ability to auto resize the NSPathControl.
the annoying thing about the previous implementation was that I
had to stretch the NSPathControl, but that made the empty space
clickable. By manually invoking sizeToFit after each set of the
path, the size becomes more manageable.

However, on extreme cases, the control will go beyond the
window. Will have to fix that later.

Signed-off-by: John Paul Alcala <ore.wa.dragon.lord@gmail.com>